pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lensing-ssc"
version = "0.2.0"
description = "A Python package for studying Super-Sample Covariance effects in weak gravitational lensing simulations"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Akira Tokiwa", email = "akira.tokiwa@example.com"},
    {name = "Adrian E. Bayer", email = "adrian.bayer@example.com"},
    {name = "Jia Liu", email = "jia.liu@example.com"},
    {name = "Masahiro Takada", email = "masahiro.takada@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Physics",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "matplotlib>=3.5.0",
    "astropy>=5.0.0",
    "pyyaml>=6.0",
    "pandas>=1.3.0",
    "h5py>=3.0.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
heavy = [
    "healpy>=1.15.0",
    "lenstools>=1.0.0",
    "nbodykit>=0.3.0",
    "classy>=3.0.0",
    "cobaya>=3.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "pre-commit>=2.20.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "numpydoc>=1.4.0",
    "jupyter-book>=0.13.0",
]

[project.urls]
Homepage = "https://github.com/atokiwaipmu/LensingSSC"
Documentation = "https://lensingssc.readthedocs.io"
Repository = "https://github.com/atokiwaipmu/LensingSSC"
Issues = "https://github.com/atokiwaipmu/LensingSSC/issues"

[project.scripts]
lensing-ssc = "lensing_ssc.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["lensing_ssc*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["lensing_ssc"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["lensing_ssc"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]